three filters for game list fixed bug in badges on game list
