+++ chess_app/templates/game_list.html
@@ -1,4 +1,5 @@
 {% extends 'base.html' %}
+{% load get_dict_value %}
 
 {% block nav-index %}
     <a class="nav-link active" href="{% url 'chess_viewer:index' %}">Lista partii </a>
@@ -16,7 +17,8 @@
                     <div class="form-row d-flex justify-content-center">
                         <div class="col-md-5 mb-3">
                             <label for="gamelastname">Nazwisko</label>
-                            <input type="text" class="form-control" id="gamelastname" value="">
+                            <input type="text" class="form-control" id="gamelastname"
+                                   value="">
 
                         </div>
                         <div class="col-md-5 mb-3">
@@ -29,28 +31,38 @@
 
                         <div class="col-md-3 mb-3">
                             <label for="tournamentdate">Data partii</label>
-                            <input type="date" class="form-control" id="tournamentdate">
-
+                            <input type="date" class="form-control" id="tournamentdate"
+                                   name="tournament__date__ym"
+                                   value="{{ filter.data | get_dict_value:"tournament__date__ym" }}">
                         </div>
                         <div class="col-md-4 mb-3">
                             <label for="tournamentcity">Nazwa turnieju</label>
-                            <input type="text" class="form-control" id="tournamentcity">
+                            <input type="text" class="form-control" id="tournamentcity"
+                                   name="tournament__name__icontains"
+                                   value="{{ filter.data | get_dict_value:"tournament__name__icontains" }}">
 
                         </div>
                         <div class="col-md-3 mb-3">
 
                             <label for="tournamenttype">Rodzaj partii</label>
-                            <select class="custom-select" id="tournamenttype">
-                                <option selected value="">...</option>
-                                <option>Klasyczna</option>
-                                <option>Szybkia</option>
-                                <option>Błyskawiczna</option>
+                            <select class="custom-select" id="tournamenttype" name="tournament__type">
+                                {% with filter_type_value=filter.data|get_dict_value:"tournament__type" %}
+                                    {% for k, v in select_options.items %}
+                                        <option value="{{ k }}"
+                                                {% if filter_type_value == k %}
+                                                selected="selected"
+                                                {% endif %}>
+                                            {{ v }}
+                                        </option>
+                                    {% endfor %}
+                                {% endwith %}
                             </select>
                         </div>
 
                     </div>
                     <div class="form-row d-flex justify-content-center">
-                        <button class="btn btn-info px-5" type="submit">Filtruj</button>
+                        <button class="btn btn-info px-5" type="submit">Filtruj</button>&nbsp;
+                        <a class="btn btn-info px-5" href="{% url 'chess_viewer:index' %}">Wyczyść</a>
                     </div>
                 </form>
             </div>
@@ -86,8 +98,10 @@
                         {{ game.tournament.date }}{% endif %}</b>
                 </div>
                 <div class="">
-                    <span class="badge badge-info">{% if tournament.type == 's' %}szachy szybkie:
-                    {% elif tournament.type == 'b' %}szachy błyskawiczne: {% else %}szachy klasyczne: {% endif %}
+                    <span class="badge badge-info">
+                        {% if game.tournament.type == 's' %}szachy szybkie:
+                        {% elif game.tournament.type == 'b' %}szachy błyskawiczne:
+                        {% else %}szachy klasyczne: {% endif %}
                         {% if game.tournament.time and game.tournament.time_add %}
                             P{{ game.tournament.time | floatformat }}'+{{ game.tournament.time_add | floatformat }}''
                         {% endif %}
@@ -96,6 +110,10 @@
             </span>
                 {% endif %}
             </a>
+        {% empty %}
+            <div class="alert alert-danger text-center" role="alert">
+                Nie znaleziono turniejów o zadanych kryteriach.
+            </div>
         {% endfor %}
 
     </div>
